Add Travis CI configuration
This is more complex than I usually do, but I wanted both
the package and the example to be tested.

Also the Makefiles are setup to be mostly silent (only print
the result of the tests themselves) and use virtual environments
to ensure that both test suites remain independent.

See https://docs.travis-ci.com/user/languages/python