Fix Travis CI configuration
Each build already runs in a virtual environment... and it doesn't
seem possible to create new virtual environments from there.

This means that both test suites won't be totally independent,
but the example (which installs dependencies) is run last, so that
should still be okay.

Also, calling unittest as a Python module seems to only find test files
when Python 3 is being used. Calling the test file directly is not as
nice, but it behaves more consitently.